+++ buCore/src/buCoreGraphicsAPI.cpp
@@ -1,0 +1,8 @@
+#include "buCoreGraphicsAPI.h"
+
+namespace buEngineSDK {
+  buCoreGraphicsAPI& g_graphicsAPI()
+  {
+    return buCoreGraphicsAPI::instance();
+  }
+}